# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Event-driven notification system assessment API.",
        "title": "Insider One Assessment Notification API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:18080",
    "basePath": "/",
    "paths": {
        "/batches": {
            "post": {
                "description": "Creates one or more notifications in a batch and stores matching outbox events for async dispatch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Create notification batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional idempotency key for safe retries",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Create notification batch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_request.CreateNotificationsRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.CreateNotificationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/batches/{id}": {
            "get": {
                "description": "Returns a notification batch with its notifications.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Get notification batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationBatchResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/notifications": {
            "get": {
                "description": "Lists notifications with filtering and pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Notification channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created at lower bound (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created at upper bound (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.ListNotificationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a single notification and stores a matching outbox event for async dispatch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional idempotency key for safe retries",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Create notification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_request.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "description": "Returns a single notification by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}/cancel": {
            "post": {
                "description": "Cancels a pending or queued notification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Cancel notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "insider-one-assessment_internal_dto_request.CreateNotificationRequest": {
            "type": "object",
            "required": [
                "notification"
            ],
            "properties": {
                "notification": {
                    "$ref": "#/definitions/insider-one-assessment_internal_dto_request.NotificationItem"
                }
            }
        },
        "insider-one-assessment_internal_dto_request.CreateNotificationsRequest": {
            "type": "object",
            "required": [
                "notifications"
            ],
            "properties": {
                "notifications": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/insider-one-assessment_internal_dto_request.NotificationItem"
                    }
                }
            }
        },
...
        "insider-one-assessment_internal_dto_request.NotificationItem": {
            "type": "object",
            "required": [
                "channel",
                "recipient"
            ],
            "properties": {
                "channel": {
                    "enum": [
                        "sms",
                        "email",
                        "push"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationChannel"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "priority": {
                    "enum": [
                        "high",
                        "normal",
                        "low"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationPriority"
                        }
                    ]
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                }
            }
        },
...
        "insider-one-assessment_internal_dto_resource.CreateNotificationsResponse": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "notification_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "replayed": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationStatus"
                }
            }
        },
        "insider-one-assessment_internal_dto_resource.ListNotificationsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationResource"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "insider-one-assessment_internal_dto_resource.NotificationBatchResource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.NotificationResource"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "insider-one-assessment_internal_dto_resource.NotificationResource": {
            "type": "object",
            "properties": {
                "attempt_count": {
                    "type": "integer"
                },
                "batch_id": {
                    "type": "string"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "channel": {
                    "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationChannel"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "max_attempts": {
                    "type": "integer"
                },
                "next_attempt_at": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationPriority"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationStatus"
                },
                "template_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "insider-one-assessment_internal_model.NotificationChannel": {
            "type": "string",
            "enum": [
                "sms",
                "email",
                "push"
            ],
            "x-enum-varnames": [
                "ChannelSMS",
                "ChannelEmail",
                "ChannelPush"
            ]
        },
        "insider-one-assessment_internal_model.NotificationPriority": {
            "type": "string",
            "enum": [
                "high",
                "normal",
                "low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityNormal",
                "PriorityLow"
            ]
        },
        "insider-one-assessment_internal_model.NotificationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "queued",
                "processing",
                "sent",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusQueued",
                "StatusProcessing",
                "StatusSent",
                "StatusFailed",
                "StatusCancelled"
            ]
        },
        "insider-one-assessment_internal_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}